[FEAT] 당첨 통계 자료 받아와서 DB에 저장하기

+++ src/main/java/com/lotto/stats/dto/Stats.java
@@ -14,6 +14,6 @@ public class Stats {
     private Integer drawNum;                                            // 회차
     private Integer lottoRank;                                          // 등수
     private Integer winnerCount;                                        // 당첨자 수
-    private Integer prizeAmount;                                        // 당첨 금액
+    private long prizeAmount;                                           // 당첨 금액
     private String crtDt;                                               // 생성 일자
 }
